+++ backend/src/main/scala/cromwell/backend/standard/StandardAsyncExecutionActor.scala
@@ -23,7 +23,7 @@ import cromwell.backend._
 import cromwell.backend.async.AsyncBackendJobExecutionActor._
 import cromwell.backend.async._
 import cromwell.backend.standard.StandardAdHocValue._
-import cromwell.backend.standard.costestimation.{AsyncJobHasFinished, ProcessThisPollResult}
+import cromwell.backend.standard.pollmonitoring.{AsyncJobHasFinished, ProcessThisPollResult}
 import cromwell.backend.standard.retry.memory.MemoryRetryResult
 import cromwell.backend.validation._
 import cromwell.core._
